requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
requests==2.32.3
shiboken6==6.8.2.1
urllib3==2.3.0
Werkzeug==3.1.3
edge-tts==6.1.9

# 语音识别相关依赖
torch==2.1.0
transformers==4.35.0
librosa==0.10.1
numpy==1.24.3
soundfile==0.12.1

# LangChain RAG 相关依赖
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.28
langchain-text-splitters==0.3.2
langchain-chroma==0.1.4
langchain-ollama==0.2.1
chromadb==0.5.23
sentence-transformers==3.3.1

# 文档处理相关依赖
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.12
unstructured==0.16.9
markdown==3.7

# 情感系统相关依赖
# 注意：这些依赖是可选的，如果不需要高级情感分析可以不安装
# transformers==4.35.0  # 用于高级情感分析（可选）
# torch==2.1.0  # transformers的依赖（可选）

# 多模态AI相关依赖
Pillow==10.4.0
opencv-python==4.10.0.84
requests==2.32.3

# 多模态AI模型相关（可选，用于高级功能）
# transformers==4.35.0  # 用于BLIP等视觉语言模型
# torch==2.1.0  # transformers依赖
# torchvision==0.16.0  # 图像处理
# clip-by-openai==1.0  # CLIP模型（如果使用）

# 语音情感合成相关依赖
pydub==0.25.1
numpy==1.24.3

# 高级音频处理（可选）
# scipy==1.11.4  # 信号处理
# librosa==0.10.1  # 音频分析
# soundfile==0.12.1  # 音频文件处理

# 高级RAG系统相关依赖
networkx==3.2.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
spacy==3.7.2

# 图谱可视化（可选）
# matplotlib==3.8.2  # 基础图表
# plotly==5.17.0  # 交互式可视化
# pyvis==0.3.2  # 网络图可视化

# 高级NLP处理（可选）
# transformers==4.35.0  # 已在多模态部分定义
# torch==2.1.0  # 已在多模态部分定义
